Move ClientTest from the verify package so it can be shared with other client tests.
